requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.2
langchain-groq==0.1.0
langchain-community==0.0.21
langgraph==0.1.0
langchain-core==0.1.17
python-dotenv==1.0.0
